requirements.txt
pika==0.10.0
python-logstash==0.4.5
